@@ -4,64 +4,62 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "ext-json": "*",
-        "api-platform/core": "^2.6",
-        "composer/package-versions-deprecated": "1.11.99.4",
-        "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^2.4",
-        "doctrine/doctrine-migrations-bundle": "^3.2",
-        "doctrine/orm": "^2.10",
-        "nelmio/cors-bundle": "^2.1",
-        "ovh/ovh": "^3.0",
+        "ext-intl": "*",
+        "api-platform/core": "^3.2",
+        "doctrine/annotations": "^2.0",
+        "doctrine/doctrine-bundle": "^2.11",
+        "doctrine/doctrine-migrations-bundle": "^3.3",
+        "doctrine/orm": "^2.17",
+        "nelmio/cors-bundle": "^2.4",
+        "ovh/ovh": "^3.2",
         "phpdocumentor/reflection-docblock": "^5.3",
-        "phpoffice/phpspreadsheet": "^1.20",
-        "sasedev/mpdf-bundle": "^2.0",
-        "sensio/framework-extra-bundle": "^6.1",
+        "phpoffice/phpspreadsheet": "^2.0",
+        "mpdf/mpdf": "^8.2",
         "symfony/apache-pack": "^1.0",
-        "symfony/asset": "5.3.*",
-        "symfony/console": "5.3.*",
-        "symfony/dotenv": "5.3.*",
-        "symfony/expression-language": "5.3.*",
-        "symfony/flex": "^1.3.1",
-        "symfony/form": "5.3.*",
-        "symfony/framework-bundle": "5.3.*",
-        "symfony/http-client": "5.3.*",
-        "symfony/intl": "5.3.*",
-        "symfony/mailer": "5.3.*",
-        "symfony/mime": "5.3.*",
-        "symfony/monolog-bundle": "^3.1",
-        "symfony/notifier": "5.3.*",
-        "symfony/process": "5.3.*",
-        "symfony/property-access": "5.3.*",
-        "symfony/property-info": "5.3.*",
-        "symfony/proxy-manager-bridge": "5.3.*",
-        "symfony/runtime": "5.3.*",
-        "symfony/security-bundle": "5.3.*",
-        "symfony/serializer": "5.3.*",
-        "symfony/string": "5.3.*",
-        "symfony/translation": "5.3.*",
-        "symfony/twig-bundle": "5.3.*",
-        "symfony/validator": "5.3.*",
-        "symfony/web-link": "5.3.*",
-        "symfony/webpack-encore-bundle": "^1.12",
-        "symfony/yaml": "5.3.*",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0",
-        "vich/uploader-bundle": "^1.18"
+        "symfony/asset": "6.4.*",
+        "symfony/console": "6.4.*",
+        "symfony/dotenv": "6.4.*",
+        "symfony/expression-language": "6.4.*",
+        "symfony/flex": "^2.4",
+        "symfony/form": "6.4.*",
+        "symfony/framework-bundle": "6.4.*",
+        "symfony/http-client": "6.4.*",
+        "symfony/intl": "6.4.*",
+        "symfony/mailer": "6.4.*",
+        "symfony/mime": "6.4.*",
+        "symfony/monolog-bundle": "^3.10",
+        "symfony/notifier": "6.4.*",
+        "symfony/process": "6.4.*",
+        "symfony/property-access": "6.4.*",
+        "symfony/property-info": "6.4.*",
+        "symfony/runtime": "6.4.*",
+        "symfony/security-bundle": "6.4.*",
+        "symfony/serializer": "6.4.*",
+        "symfony/string": "6.4.*",
+        "symfony/translation": "6.4.*",
+        "symfony/twig-bundle": "6.4.*",
+        "symfony/validator": "6.4.*",
+        "symfony/web-link": "6.4.*",
+        "symfony/webpack-encore-bundle": "^2.1",
+        "symfony/yaml": "6.4.*",
+        "twig/extra-bundle": "^3.8",
+        "twig/twig": "^3.8",
+        "vich/uploader-bundle": "^2.3"
     },
     "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^3.4",
-        "phpunit/phpunit": "^9.5",
-        "symfony/browser-kit": "5.3.*",
-        "symfony/css-selector": "5.3.*",
-        "symfony/debug-bundle": "5.3.*",
-        "symfony/maker-bundle": "^1.0",
-        "symfony/phpunit-bridge": "^5.3",
-        "symfony/stopwatch": "5.3.*",
-        "symfony/web-profiler-bundle": "5.3.*"
+        "doctrine/doctrine-fixtures-bundle": "^3.5",
+        "phpunit/phpunit": "^10.5",
+        "symfony/browser-kit": "6.4.*",
+        "symfony/css-selector": "6.4.*",
+        "symfony/debug-bundle": "6.4.*",
+        "symfony/maker-bundle": "^1.52",
+        "symfony/phpunit-bridge": "^6.4",
+        "symfony/stopwatch": "6.4.*",
+        "symfony/web-profiler-bundle": "6.4.*"
     },
     "config": {
         "optimize-autoloader": true,
