@@ -13,15 +13,11 @@ use Symfony\Component\Serializer\SerializerInterface;
 
 class UpdatePresenceController extends AbstractController
 {
-    private $em;
-    private $smsService;
-    private $serializer;
-    public function __construct(EntityManagerInterface $em, SmsService $smsService, SerializerInterface $serializer)
-    {
-        $this->em = $em;
-        $this->smsService = $smsService;
-        $this->serializer = $serializer;
-    }
+    public function __construct(
+        private readonly EntityManagerInterface $em,
+        private readonly SmsService $smsService,
+        private readonly SerializerInterface $serializer
+    ) {}
 
     public function __invoke(Request $request): Response
     {
@@ -34,20 +30,20 @@ class UpdatePresenceController extends AbstractController
                 'status' => 'error',
                 'message' => 'Invalid params'
             ], Response::HTTP_BAD_REQUEST);
-        } else {
-            $personne->setPresent($content['present']);
-            $this->em->persist($personne);
-            $this->em->flush();
-
-            $forceSms = $content['withSms'] === true && $content['present'] === true;
-
-            if ($personne->getSmsSended() !== true || $forceSms) {
-                $this->smsService->sendSms($personne);
-            }
-
-            $serializedPersonne = $this->serializer->serialize($personne, 'json');
-
-            return new JsonResponse(json_decode($serializedPersonne));
         }
+
+        $personne->setPresent($content['present']);
+        $this->em->persist($personne);
+        $this->em->flush();
+
+        $forceSms = $content['withSms'] === true && $content['present'] === true;
+
+        if ($personne->getSmsSended() !== true || $forceSms) {
+            $this->smsService->sendSms($personne);
+        }
+
+        $serializedPersonne = $this->serializer->serialize($personne, 'json');
+
+        return new JsonResponse(json_decode($serializedPersonne));
     }
 }
