@@ -2,48 +2,29 @@
 
 namespace App\Entity;
 
-use App\Controller\MediaObjectController;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\HttpFoundation\File\File;
-use Vich\UploaderBundle\Mapping\Annotation as Vich;
 use Symfony\Component\Serializer\Annotation\Groups;
+use Vich\UploaderBundle\Mapping\Annotation as Vich;
 
-/**
- * @ORM\Entity
- * @Vich\Uploadable
- */
+#[ORM\Entity]
+#[Vich\Uploadable]
 class MediaObject
 {
-    /**
-     * @var int|null
-     *
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue
-     * @ORM\Id
-     * @Groups({"admin"})
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: 'integer')]
+    #[Groups(['admin'])]
+    private ?int $id = null;
 
-    /**
-     * @var string|null
-     *
-     */
-    public $contentUrl;
+    public ?string $contentUrl = null;
 
-    /**
-     * @var File|null
-     *
-     * @Vich\UploadableField(mapping="media_object", fileNameProperty="filePath")
-     */
-    public $file;
+    #[Vich\UploadableField(mapping: 'media_object', fileNameProperty: 'filePath')]
+    public ?File $file = null;
 
-    /**
-     * @var string|null
-     *
-     * @ORM\Column(nullable=true)
-     * @Groups({"admin"})
-     */
-    public $filePath;
+    #[ORM\Column(nullable: true)]
+    #[Groups(['admin'])]
+    public ?string $filePath = null;
 
     public function getId(): ?int
     {
