@@ -2,45 +2,35 @@
 
 namespace App\Entity;
 
-use ApiPlatform\Core\Annotation\ApiResource;
+use ApiPlatform\Metadata\ApiResource;
 use App\Repository\TicketRepository;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Serializer\Annotation\Groups;
 
-/**
- * @ApiResource(
- *  normalizationContext={"groups"={"admin", "ticket"}}
- * )
- * @ORM\Entity(repositoryClass=TicketRepository::class)
- */
+#[ORM\Entity(repositoryClass: TicketRepository::class)]
+#[ApiResource(
+    normalizationContext: ['groups' => ['admin', 'ticket']]
+)]
 class Ticket
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     * @Groups({"ticket", "personne"})
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: 'integer')]
+    #[Groups(['ticket', 'personne'])]
+    private ?int $id = null;
 
-    /**
-     * @ORM\Column(type="string", length=255)
-     * @Groups({"ticket", "personne"})
-     */
-    private $fichier;
+    #[ORM\Column(type: 'string', length: 255)]
+    #[Groups(['ticket', 'personne'])]
+    private ?string $fichier = null;
 
-    /**
-     * @ORM\Column(type="integer")
-     * @Groups({"ticket", "personne"})
-     */
-    private $numero;
+    #[ORM\Column(type: 'integer')]
+    #[Groups(['ticket', 'personne'])]
+    private ?int $numero = null;
 
-    /**
-     * @ORM\OneToOne(targetEntity=Personne::class, inversedBy="ticket", cascade={"persist", "remove"}, fetch="EAGER")
-     * @ORM\JoinColumn(nullable=false)
-     * @Groups({"ticket"})
-     */
-    private $personne;
+    #[ORM\OneToOne(targetEntity: Personne::class, inversedBy: 'ticket', cascade: ['persist', 'remove'], fetch: 'EAGER')]
+    #[ORM\JoinColumn(nullable: false)]
+    #[Groups(['ticket'])]
+    private ?Personne $personne = null;
 
     public function getId(): ?int
     {
