@@ -3,25 +3,21 @@
 namespace App\Normalizer;
 
 use App\Entity\MediaObject;
-use Symfony\Component\Serializer\Normalizer\ContextAwareNormalizerInterface;
-use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
-use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
+use Symfony\Component\DependencyInjection\Attribute\Autowire;
+use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Vich\UploaderBundle\Storage\StorageInterface;
 
-final class MediaObjectNormalizer implements ContextAwareNormalizerInterface, NormalizerAwareInterface
+final class MediaObjectNormalizer implements NormalizerInterface
 {
-    use NormalizerAwareTrait;
-
     private const ALREADY_CALLED = 'MEDIA_OBJECT_NORMALIZER_ALREADY_CALLED';
 
-    private $storage;
+    public function __construct(
+        private readonly StorageInterface $storage,
+        #[Autowire(service: 'serializer.normalizer.object')]
+        private readonly NormalizerInterface $normalizer
+    ) {}
 
-    public function __construct(StorageInterface $storage)
-    {
-        $this->storage = $storage;
-    }
-
-    public function normalize($object, ?string $format = null, array $context = [])
+    public function normalize(mixed $object, ?string $format = null, array $context = []): array|string|int|float|bool|\ArrayObject|null
     {
         $context[self::ALREADY_CALLED] = true;
         $object->contentUrl = $this->storage->resolveUri($object, 'file');
@@ -29,12 +25,19 @@ final class MediaObjectNormalizer implements ContextAwareNormalizerInterface, No
         return $this->normalizer->normalize($object, $format, $context);
     }
 
-    public function supportsNormalization($data, ?string $format = null, array $context = []): bool
+    public function supportsNormalization(mixed $data, ?string $format = null, array $context = []): bool
     {
         if (isset($context[self::ALREADY_CALLED])) {
             return false;
         }
 
         return $data instanceof MediaObject;
+    }
+
+    public function getSupportedTypes(?string $format): array
+    {
+        return [
+            MediaObject::class => true,
+        ];
     }
 }
