@@ -2,20 +2,16 @@
 
 namespace App\Validator\Constraints;
 
+use Attribute;
 use Symfony\Component\Validator\Constraint;
 
-/**
- * @Annotation
- */
+#[Attribute(Attribute::TARGET_PROPERTY | Attribute::TARGET_METHOD)]
 class moyenReglementConstraint extends Constraint
 {
     public string $message = "Le moyen de paiement ne peut pas être vide";
 
-    /**
-     * @return string
-     */
     public function validatedBy(): string
     {
-        return \get_class($this) . 'Validator';
+        return static::class . 'Validator';
     }
 }
